API/requirements.txt
fastapi
uvicorn[standard]
tensorflow
Pillow
numpy
python-dotenv
google-generativeai
python-multipart
